[WRFE-55](refactor): 변경된 폴더구조로 Import 경로 수정
