+++ apps/front/wraffle-webview/src/shared/ui/bottom/index.ts
@@ -1,0 +1,1 @@
+export * from './BottomFixedBox';
