+++ apps/front/wraffle-webview/src/shared/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './bottom';
 export * from './form';
 export * from './header/Header';
 export * from './progress/Progress';
